+++ scripts/clone_plugins.sh
@@ -34,6 +34,7 @@ repos=(
   "https://github.com/huchenlei/ComfyUI-layerdiffuse 151f746"
   "https://github.com/kijai/ComfyUI-KJNodes 48d5a18"
   "https://github.com/Fannovel16/ComfyUI-Frame-Interpolation 5e11679"
+  "https://github.com/fofr/ComfyUI-HyperSDXL1StepUnetScheduler 5073c7a"
 )
 
 # Destination directory